refactor: share expo config in TS module
